feat: Add geminiZeroShotScoring

+++ src/main/java/org/ui/thesis/dtos/AiScoringResult.java
@@ -38,5 +38,5 @@ public class AiScoringResult {
 
     private String aiFeedback;
 
-    private Long tokenUsage;
+    private Integer tokenUsage;
 }
